Feat : BoxComment 분리 , Title 과 상세 설명 데이터와 연결
